agent-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain==0.3.11
langchain-core==0.3.63
langchain-openai==0.2.0
langchain-community==0.3.11
langgraph==0.2.20
duckduckgo-search